refactor: 패키지 및 클래스명 수정
- 프로젝트 통일성을 갖추기 위한 이름 수정
