+++ src/main/java/com/ssginc/wms/incomeApply/AdminIncomeApplyUI.java
@@ -1,8 +1,10 @@
-package com.ssginc.wms.supply;
+package com.ssginc.wms.incomeApply;
 
 import com.ssginc.wms.frame.AdminFrame;
-import com.ssginc.wms.incomeApply.IncomeApplyService;
 import com.ssginc.wms.product.ProductService;
+import com.ssginc.wms.supply.IncomeApplyDAO;
+import com.ssginc.wms.supply.ProductIncomeApplyVO;
+import com.ssginc.wms.supply.SupplyVO;
 
 import javax.swing.*;
 import javax.swing.table.DefaultTableModel;
@@ -17,7 +19,7 @@ public class AdminIncomeApplyUI extends AdminFrame {
     private JTable applicationTable;
     private JComboBox<String> categoryComboBox;
     private DefaultTableModel tableModel;
-    private IncomeApplyDAO incomeApplyDAO;
+    private com.ssginc.wms.supply.IncomeApplyDAO incomeApplyDAO;
     Color color = new Color(0x615959);
 
     public AdminIncomeApplyUI(String id) {
@@ -114,8 +116,8 @@ public class AdminIncomeApplyUI extends AdminFrame {
 
         allApplyButton.addActionListener(e -> {
             tableModel.setRowCount(0); // 기존 데이터 삭제
-            List<ProductIncomeApplyVO> applications = incomeApplyDAO.listIncomeApply(null, null);
-            for (ProductIncomeApplyVO application : applications) {
+            List<com.ssginc.wms.supply.ProductIncomeApplyVO> applications = incomeApplyDAO.listIncomeApply(null, null);
+            for (com.ssginc.wms.supply.ProductIncomeApplyVO application : applications) {
                 tableModel.addRow(new Object[]{
                         IncomeApplyService.encodeApplyId(application.getApplyId()),
                         ProductService.encodeProductId(application.getProductId()),
@@ -129,8 +131,8 @@ public class AdminIncomeApplyUI extends AdminFrame {
         });
         yetApplyButton.addActionListener(e -> {
             tableModel.setRowCount(0); // 기존 데이터 삭제
-            List<ProductIncomeApplyVO> applications = incomeApplyDAO.listPendingIncomeApplies();
-            for (ProductIncomeApplyVO application : applications) {
+            List<com.ssginc.wms.supply.ProductIncomeApplyVO> applications = incomeApplyDAO.listPendingIncomeApplies();
+            for (com.ssginc.wms.supply.ProductIncomeApplyVO application : applications) {
                 tableModel.addRow(new Object[]{
                         IncomeApplyService.encodeApplyId(application.getApplyId()),
                         ProductService.encodeProductId(application.getProductId()),
@@ -154,7 +156,7 @@ public class AdminIncomeApplyUI extends AdminFrame {
     public void loadProductData(String selectedColumn, String searchKeyword) {
         tableModel.setRowCount(0); // 기존 데이터 삭제
 
-        List<ProductIncomeApplyVO> applications = incomeApplyDAO.listIncomeApply(selectedColumn, searchKeyword);
+        List<com.ssginc.wms.supply.ProductIncomeApplyVO> applications = incomeApplyDAO.listIncomeApply(selectedColumn, searchKeyword);
 
         for (ProductIncomeApplyVO application : applications) {
             tableModel.addRow(new Object[]{
